Allow building from the beginning in a non master branch
NOTE: This is a bug discovered and fixed in the update-to-pyro branch.
It addresses issues that would pop up during the official releases
which can't wait the pyro update. That is why it is cherry-picked from
the original update-to-pyro branch.

It is tagged as the 0.1.5. but it doesn't influence any changes
senic-hub.

This change makes the repo syncs work correctly. Without it
it would default to the manifest file in master.